feat: 예약 대기 자동 승인 기능 추가

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -1,6 +1,7 @@
 package roomescape.reservation.service;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import roomescape.member.model.Member;
 import roomescape.member.repository.MemberRepository;
@@ -86,8 +87,17 @@ public class ReservationService {
         }
     }
 
+    @Transactional
     public void deleteReservation(final Long reservationId) {
-        reservationRepository.deleteById(reservationId);
+        if (!waitingRepository.existsByReservationId(reservationId)) {
+            reservationRepository.deleteById(reservationId);
+            return;
+        }
+        //기존 예약의 멤버를 대기 1번 멤버로 바꿔줘야함.
+        Reservation reservation = reservationRepository.findById(reservationId).get();
+        Member firstCandidate = waitingRepository.findFirstMemberByReservationIdOrderByIdAsc(reservationId);
+        reservation.setMember(firstCandidate);
+        waitingRepository.deleteByMemberAndReservation(firstCandidate, reservation);
     }
 
     public List<MyReservationResponse> getMyReservations(final Long memberId) {
